LayoutTests/media/modern-media-controls/status-label/status-label-text-selection-expected.txt

Testing that text selection is turned off for StatusLabel.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.bottomControlsBar.element == mediaControls.statusLabel.element.parentElement is true
PASS window.getComputedStyle(mediaControls.bottomControlsBar.element).webkitUserSelect is "none"
PASS window.getComputedStyle(mediaControls.statusLabel.element).cursor is "default"

PASS successfullyParsed is true

TEST COMPLETE
